Add fetching api functionality
