+++ .eslintrc.json
@@ -13,9 +13,18 @@
   "extends": ["airbnb-base"],
   "rules": {
     "no-shadow": "off",
-    "no-param-reassign": 0,
-    "eol-last": "off"
-  },
+    "no-param-reassign": "off",
+    "no-nested-ternary": "off",
+    "eol-last": "off",
+    "import/prefer-default-export":"off" ,
+    "import/extensions": [
+      1,
+      {
+        "js": "always",
+        "json": "always"
+      }
+    ]
+  }, 
 
   "ignorePatterns": ["dist/", "build/"]
 }
